+++ scripts/extensions/availability-manager/src/settings/availability-settings.tsx
